Tests/LibWeb/Text/expected/anchor-positioning-inside-outside.txt

anchor: left=100 top=50 right=300 bottom=130
inside (left/top):
left: 100 top: 50
outside (left/top):
left: 300 top: 130
inside (right/bottom):
right: 300 bottom: 130
outside (right/bottom):
right: 100 bottom: 50
